SearchScreen 디자인 추가
